[update]画像からテキスト抽出機能のJavaScriptを独立させた
画像アップロードとテキスト抽出の処理を`image_to_text.js`に分離し、`image_to_text_and_generate_questions.js`から該当するコードを削除しました。application.js`に新しいファイルのインポートを追加し、`importmap.rb`にもこの変更を反映させ、依存関係を明確にしました。
