+++ app/javascript/application.js
@@ -12,3 +12,4 @@ import "hamburger_menu"
 import "display_answer"
 import "quiz_details"
 import "toggle_publish"
+import "image_to_text"
