+++ config/importmap.rb
@@ -12,3 +12,4 @@ pin "hamburger_menu", to: "hamburger_menu.js", preload: true
 pin "display_answer", to: "display_answer.js", preload: true
 pin "quiz_details", to: "quiz_details.js", preload: true
 pin "toggle_publish", to: "toggle_publish.js", preload: true
+pin "image_to_text", to: "image_to_text.js", preload: true
